fix: allow end of string as last character of file
Fixes issue #61

If `allowInvalid` is set to false, we threw an exception if the last
character of the CSV was a closing text-delimiter.

We didn't recognize this, as there is the possibility, that another
text-delimiter "uncloses" a text-field.  This is of course not possible
if there are no characters left.